Move assistant and reporting modules into packages
