+++ control-plane/aegisops_control_plane/reporting/__init__.py
@@ -1,1 +1,1 @@
-"""Package scaffold only for future reporting and inspection modules."""
+"""Package scaffold only; reporting domain modules are owned here."""
