+++ control-plane/aegisops_control_plane/service.py
@@ -13,14 +13,14 @@ from typing import Iterable, Iterator, Mapping, Protocol, Type, TypeVar
 _DATETIME_TYPE = datetime
 
 from .action_policy import evaluate_action_policy_record
-from . import assistant_advisory as _assistant_advisory
-from . import live_assistant_workflow as _live_assistant_workflow
-from .assistant_provider import (
+from .assistant import assistant_advisory as _assistant_advisory
+from .assistant import live_assistant_workflow as _live_assistant_workflow
+from .assistant.assistant_provider import (
     AssistantProviderFailure,
     AssistantProviderResult,
     AssistantProviderTransport,
 )
-from .audit_export import export_audit_retention_baseline
+from .reporting.audit_export import export_audit_retention_baseline
 from .config import OpenBaoKVv2SecretTransport, RuntimeConfig
 from .models import (
     AITraceRecord,
